finally fixed octals, hex should be quick and moving on to f and p

+++ libftprintf/ft_atoi.c
@@ -6,15 +6,15 @@
 /*   By: nkellum <nkellum@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/13 14:59:06 by nkellum           #+#    #+#             */
-/*   Updated: 2019/03/11 14:17:20 by nkellum          ###   ########.fr       */
+/*   Updated: 2019/03/27 20:33:32 by nkellum          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libftprintf.h"
 
-int	ft_atoi(const char *str)
+long long ft_atoi(const char *str)
 {
-	int num;
+	long long num;
 	int isneg;
 	int i;
